infrastructure/functions/rotate-django-secret/requirements.txt
functions-framework==3.*
google-cloud-secret-manager==2.*
cloudevents==1.*
